+++ kedro_kubeflow/context_helper.py
@@ -5,7 +5,6 @@ from kedro import __version__ as kedro_version
 from semver import VersionInfo
 
 from .config import PluginConfig
-from .kfpclient import KubeflowClient
 
 
 class ContextHelper(object):
@@ -16,31 +15,24 @@ class ContextHelper(object):
         self,
         metadata,
         env,
-        username,
-        password,
-        experiment_namespace,
-        namespace,
     ):
         self._metadata = metadata
         self._env = env
-        self._username = username
-        self._password = password
-        self._experiment_namespace = experiment_namespace
-        self._namespace = namespace
 
     @property
     def project_name(self):
         return self._metadata.project_name
 
     @property
-    def context(self):
+    @lru_cache()
+    def session(self):
         from kedro.framework.session import KedroSession
 
-        return KedroSession.create(
-            self._metadata.package_name,
-            env=self._env,
-            extra_params={},
-        ).load_context()
+        return KedroSession.create(self._metadata.package_name, env=self._env)
+
+    @property
+    def context(self):
+        return self.session.load_context()
 
     @property
     @lru_cache()
@@ -51,42 +43,36 @@ class ContextHelper(object):
     @property
     @lru_cache()
     def kfp_client(self):
-        return KubeflowClient(
-            self.config,
-            self.project_name,
-            self.context,
-            self._username,
-            self._password,
-            self._namespace,
-        )
+        if self.config.is_vertex_ai_pipelines:
+            from .vertex_ai.client import VertexAIPipelinesClient
+
+            return VertexAIPipelinesClient(
+                self.config, self.project_name, self.context
+            )
+        else:
+            from .kfpclient import KubeflowClient
+
+            return KubeflowClient(
+                self.config,
+                self.project_name,
+                self.context,
+            )
 
     @staticmethod
     def init(
         metadata,
         env,
-        username,
-        password,
-        experiment_namespace,
-        namespace,
     ):
         version = VersionInfo.parse(kedro_version)
         if version.match(">=0.17.0"):
             return ContextHelper(
                 metadata,
                 env,
-                username,
-                password,
-                experiment_namespace,
-                namespace,
             )
         else:
             return ContextHelper16(
                 metadata,
                 env,
-                username,
-                password,
-                experiment_namespace,
-                namespace,
             )