+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+-   Support for kedro-mlflow>=0.7
+-   Support of Google Vertex AI Pipelines (EXPERIMENTAL)
+
 ## [0.3.1] - 2021-05-25
 
 ### Fixed